Fixed jittery image

+++ src/cs3500/animation/model/Transformation.java
@@ -32,15 +32,15 @@ public class Transformation {
    * @param g1   initial green color
    * @param b1   initial blue color
    */
-  public Transformation(String name, int t1, int x1, int y1,
+  public Transformation(String name, int t1, double x1, double y1,
                         int w1, int h1, int r1, int g1, int b1) {
     this.name = name;
     this.t1 = t1;
-    this.posn1 = new Position2D(x1, y1);
+    this.posn1 = new Position2D((int) x1,(int) y1);
     this.dimn1 = new Dimension2D(w1, h1);
     this.color1 = new Color(r1, g1, b1);
     this.t2 = t1;
-    this.posn2 = new Position2D(x1, y1);
+    this.posn2 = new Position2D((int) x1, (int) y1);
     this.dimn2 = new Dimension2D(w1, h1);
     this.color2 = new Color(r1, g1, b1);
   }
@@ -66,16 +66,16 @@ public class Transformation {
    * @param g2   final green color
    * @param b2   final blue color
    */
-  Transformation(String name, int t1, int x1, int y1, int w1, int h1, int r1, int g1, int b1,
+  Transformation(String name, int t1, double x1, double y1, int w1, int h1, int r1, int g1, int b1,
                  int t2,
-                 int x2, int y2, int w2, int h2, int r2, int g2, int b2) {
+                 double x2, double y2, int w2, int h2, int r2, int g2, int b2) {
     this.name = name;
     this.t1 = t1;
-    this.posn1 = new Position2D(x1, y1);
+    this.posn1 = new Position2D((int) x1, (int) y1);
     this.dimn1 = new Dimension2D(w1, h1);
     this.color1 = new Color(r1, g1, b1);
     this.t2 = t2;
-    this.posn2 = new Position2D(x2, y2);
+    this.posn2 = new Position2D((int) x2, (int)y2);
     this.dimn2 = new Dimension2D(w2, h2);
     this.color2 = new Color(r2, g2, b2);
   }
